+++ src/pages/embed.tsx
@@ -7,8 +7,8 @@ import {
   HighlightValue,
 } from "src/components/highlight-context";
 import { MapProvider } from "src/components/map-context";
-import { thresholdEncodings } from "src/domain/charts";
 import { PriceComponent } from "src/domain/data";
+import { thresholdEncodings } from "src/domain/map-encodings";
 import { useQueryStateEnergyPricesMap } from "src/domain/query-states";
 import { PriceComponent as PriceComponentEnum } from "src/graphql/queries";
 import { useEnrichedEnergyPricesData } from "src/hooks/use-enriched-energy-prices-data";
@@ -59,7 +59,11 @@ const IndexPage = ({ locale }: Props) => {
   const colorAccessor = useCallback((d: { value: number }) => d.value, []);
   const colorScale = useMemo(() => {
     const values = observations.filter(isValidValue).map(colorAccessor);
-    const encoding = thresholdEncodings.energyPrices(medianValue, values, +period);
+    const encoding = thresholdEncodings.energyPrices(
+      medianValue,
+      values,
+      +period
+    );
     return encoding.makeScale();
   }, [colorAccessor, medianValue, observations, period]);